moedulizing add client component
